✨ Added authorization header in App.vue and added clear functionality in Datable.vue and Search.vue
The commit adds an authorization header to the headers array in App.vue. In Datable.vue, a new event listener `@clear` is added to the `<datable-search>` component, which triggers the `clearSearch` method. Similarly, in Search.vue, a new event listener `@clear` is added to the emit function, triggering the `clear` method. This allows for clearing of search queries when triggered.